fix(build): cast Long timeout to Int at ExtractorApi.kt:161
NiceHTTP.head() now takes timeout:Int (matching real lib JVM descriptor).
ExtractorApi.getVideoSize() passes timeoutSeconds:Long — add .toInt() cast.

+++ StreamCloud/app/src/main/java/com/lagradost/cloudstream3/utils/ExtractorApi.kt
@@ -158,7 +158,7 @@ constructor(
     suspend fun getVideoSize(timeoutSeconds: Long = 3L): Long? {
         if (type != ExtractorLinkType.VIDEO) return null
         return try {
-            app.head(url, headers = headers, referer = referer, timeout = timeoutSeconds)
+            app.head(url, headers = headers, referer = referer, timeout = timeoutSeconds.toInt())
                 .headers["Content-Length"]?.firstOrNull()?.toLong()
         } catch (_: Exception) { null }
     }
